Drop Relation.Internal(Derived), Symmetric.Internal
