+++ src/Algebra/Graph/Relation/Transitive.hs
@@ -1,7 +1,7 @@
 -----------------------------------------------------------------------------
 -- |
 -- Module     : Algebra.Graph.Relation.Transitive
--- Copyright  : (c) Andrey Mokhov 2016-2018
+-- Copyright  : (c) Andrey Mokhov 2016-2019
 -- License    : MIT (see the file LICENSE)
 -- Maintainer : andrey.mokhov@gmail.com
 -- Stability  : experimental
@@ -14,8 +14,45 @@ module Algebra.Graph.Relation.Transitive (
     TransitiveRelation, fromRelation, toRelation
     ) where
 
+import Control.DeepSeq
 import Algebra.Graph.Relation
-import Algebra.Graph.Relation.InternalDerived
+
+import qualified Algebra.Graph.Class as C
+
+-- TODO: Optimise the implementation by caching the results of transitive closure.
+{-| The 'TransitiveRelation' data type represents a /transitive binary relation/
+over a set of elements. Transitive relations satisfy all laws of the
+'Transitive' type class and, in particular, the /closure/ axiom:
+
+@y /= 'empty' ==> x * y + x * z + y * z == x * y + y * z@
+
+For example, the following holds:
+
+@'path' xs == ('clique' xs :: TransitiveRelation Int)@
+
+The 'Show' instance produces transitively closed expressions:
+
+@show (1 * 2         :: TransitiveRelation Int) == "edge 1 2"
+show (1 * 2 + 2 * 3 :: TransitiveRelation Int) == "edges [(1,2),(1,3),(2,3)]"@
+-}
+newtype TransitiveRelation a = TransitiveRelation { fromTransitive :: Relation a }
+    deriving (Num, NFData)
+
+instance Ord a => Eq (TransitiveRelation a) where
+    x == y = transitiveClosure (fromTransitive x) == transitiveClosure (fromTransitive y)
+
+instance (Ord a, Show a) => Show (TransitiveRelation a) where
+    show = show . transitiveClosure . fromTransitive
+
+-- TODO: To be derived automatically using GeneralizedNewtypeDeriving in GHC 8.2
+instance Ord a => C.Graph (TransitiveRelation a) where
+    type Vertex (TransitiveRelation a) = a
+    empty       = TransitiveRelation empty
+    vertex      = TransitiveRelation . vertex
+    overlay x y = TransitiveRelation $ fromTransitive x `overlay` fromTransitive y
+    connect x y = TransitiveRelation $ fromTransitive x `connect` fromTransitive y
+
+instance Ord a => C.Transitive (TransitiveRelation a)
 
 -- | Construct a transitive relation from a 'Relation'.
 -- Complexity: /O(1)/ time.
