+++ test/Algebra/Graph/Test/API.hs
@@ -38,9 +38,7 @@ import qualified Algebra.Graph.AdjacencyMap.Algorithm         as AM
 import qualified Algebra.Graph.Labelled                       as LG
 import qualified Algebra.Graph.Labelled.AdjacencyMap          as LAM
 import qualified Algebra.Graph.Relation                       as R
-import qualified Algebra.Graph.Relation.Internal              as R
 import qualified Algebra.Graph.Relation.Symmetric             as SR
-import qualified Algebra.Graph.Relation.Symmetric.Internal    as SR
 import qualified Algebra.Graph.ToGraph                        as T
 
 import Algebra.Graph.Test.Arbitrary ()
